Convert Modal State Manager into Custom Hook
Using this custom hook, we can reuse this modal elsewhere (and even outside this project if we use styled components).

The hook returns the current values, ways to mutate them, and a ref to connect a placed Modal (so that it can sync its visibility outside of that which is managed by the Modal itself).