Refactor rearrange_poi into just rearrange
Since rearranging the POI and the day are almost the same thing (just POI uses the FK to categorize the POIs belonging to that day), we refactor the reducer to contain only a case for rearrange.

Then, the payload should give a 'type' and 'fk' attribute which will help identify the conflicting record, and then swap those values.