Modify add case logic for reducer
Now we take order into account when adding records to an array. If the order property exists on the payload, then we determine if the array already possesses an item with that same order.  If so, we splice it there and  then we use a for loop to increment each value in the array thereafter. If not, we add it to the end.