Add logic for loading a placeholder when adding an image during ADD_POI
A placeholder is now shown whilst the user waits for the image to upload to the server and for their client to download it for them.

The dispatch adds the placeholder while putting the real path in 'realpath'. In the afterware section, after the image is uploaded, the photo is then replaced via dispatch.

A use-effect hook watches for any changes to photos during POIDetails' renderings, of which it then refreshes its photos state variable, which leads to the file being shown.