Fix issue with undefined day in Add POI
The problem was giving the selDay state variable the entire activeDay object instead of just the data property.

Then we replaced the get_greatest_order logic to use a simple sort and we pick the last element from this sort.

One thing that really helps is using console.log to describe where we are in the call stack since there are 50 other log messages that obfuscate what we're trying to debug. (and it doesn't help that React double-renders, spamming each message twice.)