Implement ability to delete POIs!
Had some issues with promises, but the key understanding was that they are executed as soon as they are created.

Thus, I opted instead to BIND the async functions and then call them in batches using splice, thus delaying their execution until their time of actual creation (when invoking the function).