Refactor delete method to ensure multiple deletes
The problem was that before, multiple deletes couldn't occur due to:
1. mutations causing exceptions in the firebase module.
2. for some reason, the second delete operation would never fire no matter what. I ended up figuring out that the getDoc operation would screw up the asynchronous chain.

So I removed the getDoc operation from delete since we just need the storageUri location -- and then I include directly on the object on load.

Avoiding that operation miraculously solves the problems with multiple delete. Hm, I suppose in the future we would've seen an exception message with some catches anywhere we had an async operation.