Remove lengthy effect hook for updating sidebar content
Turns out the solution we had before was just fine. No need for conditionals in a single hook; it's much more neater to write this in two blocks.

And before you go wondering if they will be conflicts -- the user can only interact with one of those components at a time, which means they can't have the other just show up (it wouldn't run such logic since the reference to the object has not changed)