Fix unhandled mutation error from Firestore update
The problem is with STALE DOCUMENT REFERENCES... each time they're changed, seems like something internally with the object changes, which means you need to get a new reference for the document.

If we see this occur again, we'll need to replace the use of *.ref with getRef and the ref objects provided path.