Add styling for the gallery view (display mode)
The header and the view is styled and responsive to width and height changes. The image is set to be its original upload size, but this is limited if this is greater than the viewport's width (100%) and about half of the viewport's height. The next part to style is the modal that is used in this component.