Fix occasionally occurring double POI click issue
This is because this effect hook is running to update the activeDay each time the activePin changes.

This triggers the effect hook t o show the activeDay once again, hence the user needing to double click POIs to access them.

I imagine the reference is recreated, causing React to want to re-render again when it performs a referential check.