Fix bug with add poi search marker still appearing despite closing details sidebar
The solution was to use a mutation observer which observes the sidebar for changes in its attributes. If at any point whilst the location input component is active, and we close the sidebar, the observer invokes a func. which verifies that the display status is none. If so, it removes the search marker from view.