Add mobile fix -- unfold sidebars when searching location as not to block map
If we determine the user is browsing on a device smaller than a tablet, then the two sidebars will take up the width of the screen and block the map. Thus, we use the query selector to simulate clicking both of those buttons so the user can see where the pin has landed on the map.